+++ jdaviz/core/viewer_creators/histogram/histogram.py
@@ -1,6 +1,6 @@
 from traitlets import List, Unicode, observe
 
-from jdaviz.configs.default.plugins.viewers import ScatterViewer
+from jdaviz.configs.default.plugins.viewers import HistogramViewer
 from jdaviz.core.user_api import ViewerCreatorUserApi
 from jdaviz.core.viewer_creators import BaseViewerCreator
 from jdaviz.core.registries import viewer_creator_registry
@@ -8,39 +8,31 @@ from jdaviz.core.template_mixin import SelectPluginComponent
 from jdaviz.utils import att_to_componentid
 
 
-__all__ = ['ScatterViewerCreator']
+__all__ = ['HistogramViewerCreator']
 
 
-@viewer_creator_registry('Scatter', overwrite=True)
-class ScatterViewerCreator(BaseViewerCreator):
-    template_file = __file__, "scatter.vue"
-
+@viewer_creator_registry('Histogram', overwrite=True)
+class HistogramViewerCreator(BaseViewerCreator):
+    template_file = __file__, "histogram.vue"
     xatt_items = List().tag(sync=True)
     xatt_selected = Unicode().tag(sync=True)
 
-    yatt_items = List().tag(sync=True)
-    yatt_selected = Unicode().tag(sync=True)
-
     def __init__(self, app, **kwargs):
         super().__init__(app, **kwargs)
-        self.viewer_label_default = 'Scatter'
+        self.viewer_label_default = 'Histogram'
 
         self.xatt = SelectPluginComponent(self,
                                           items='xatt_items',
                                           selected='xatt_selected',
                                           default_mode='first')
-        self.yatt = SelectPluginComponent(self,
-                                          items='yatt_items',
-                                          selected='yatt_selected',
-                                          default_mode='second')
 
     @property
     def user_api(self):
-        return ViewerCreatorUserApi(self, expose=('xatt', 'yatt'))
+        return ViewerCreatorUserApi(self, expose=('xatt',))
 
     @property
     def viewer_class(self):
-        return ScatterViewer
+        return HistogramViewer
 
     @observe('dataset_selected')
     def _on_dataset_selected_change(self, event):
@@ -56,16 +48,12 @@ class ScatterViewerCreator(BaseViewerCreator):
         valid_atts = sorted(set(valid_atts))
         self.xatt._manual_options = valid_atts
         self.xatt._update_items()
-        self.yatt._manual_options = valid_atts
-        self.yatt._update_items()
 
     def __call__(self):
         nv = super().__call__()
         gv_state = nv._obj.glue_viewer.state
         if self.xatt.selected != '':
             gv_state.x_att = att_to_componentid(gv_state.x_att_helper, self.xatt.selected)
-        if self.yatt.selected != '':
-            gv_state.y_att = att_to_componentid(gv_state.y_att_helper, self.yatt.selected)
-        if self.xatt.selected != '' or self.yatt.selected != '':
+        if len(nv._obj.glue_viewer.layers):
             nv._obj.glue_viewer.reset_limits()
         return nv